Add atom to set / refresh product list
